Add  js folder
