+++ js/modules/dom.js
@@ -1,9 +1,10 @@
+/* eslint-disable import/no-cycle */
+import fetchData from '../fuctionalties.js';
+
 const API_KEY = 's5I1Iezs2fp5U1vv55v4';
 export const URI = `https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/${API_KEY}/scores`;
-
-
 export const displayScores = (list, res) => {
-  const { result } = undefined;
+  const { result } = res;
   const scores = [];
   for (let i = 0; i < result.length; i += 1) {
     scores.push([result[i].user, result[i].score]);
