+++ src/index.js
@@ -1,20 +1,10 @@
 // eslint-disable-next-line no-unused-vars
-import { URI, displayScores } from '../modules/dom.js';
-import { sendData, postData } from '../modules/refresh.js';
+import { URI, displayScores } from '../js/modules/dom.js';
+
+import sendData from '../js/modules/refresh.js';
+import postData from '../js/fuctionalties.js';
 
 displayScores();
+
+sendData();
 postData();
-const resultUl = document.querySelector('.board-list');
-
-const fetchData = async () => {
-  await fetch(URI, {
-    method: 'GET',
-  })
-    .then((response) => response.json())
-    .then((result) => displayScores(resultUl, result));
-};
-
-document.addEventListener('DOMContentLoaded', () => {
-  sendData();
-  fetchData();
-});
